Added ansible module to collect openstack-release information.
This is in effort to run openstack ansible in liberty, mitaka,
and newton+ environments. In mitaka and older, openstack ansible
needs to know the location of the openstack ansible inventory
if not run from the the openstack ansible playbook directory.

This change allows us to detect the major version of the openstack
ansible release and branch accordingly.